fix: ambulance state missing puedeAtenderDomicilio added, solicitante added as param to every method
