--- src/PatronStateAmbulancia/EstadoAbstracto.java
@@ -1,10 +1,0 @@
-package PatronStateAmbulancia;
-
-public abstract class EstadoAbstracto implements IEstadoAmbulancia{
-    protected Ambulancia ambulancia;
-    public EstadoAbstracto(Ambulancia ambulancia){
-        this.ambulancia = ambulancia;
-    }
-    public boolean puedeTraslado(){ return false; }
-    public boolean puedeMantenimiento(){ return false; }
-}
