+++ src/mdp/ingenieria/clinicagestion/model/ambulancia/EstadoTrasladando.java
@@ -1,10 +1,9 @@
-package PatronStateAmbulancia;
+package mdp.ingenieria.clinicagestion.model.ambulancia;
 
-public class EstadoAtendiendo extends EstadoAbstracto{
-    public EstadoAtendiendo(Ambulancia ambulancia) {
+public class EstadoTrasladando extends EstadoAbstracto{
+    public EstadoTrasladando(Ambulancia ambulancia) {
         super(ambulancia);
     }
-
     @Override
     public void solicitarAtencionDomicilio() {
         // permanece en este estado
@@ -14,13 +13,14 @@ public class EstadoAtendiendo extends EstadoAbstracto{
     public void solicitarTraslado() {
         // informa que no puede
     }
+
     @Override
     public void retornoAutomatico() {
-        this.ambulancia.setEstado(new EstadoRegresando(this.ambulancia));
+        this.ambulancia.setEstado(new EstadoDisponible(this.ambulancia));
     }
 
     @Override
     public void solicitarMantenimiento() {
-        // informa que no puede
+        // infroma que no puede
     }
 }
